refactor: move river tests to tests directory
Moved the embedded unit tests from `src/mahjong/river.rs` to a new integration test file at `tests/test_river.rs` to adhere to the project's testing conventions. The new file properly imports the required structs and all tests pass successfully.
